requirements.txt
Flask
eventlet
Flask-SocketIO==5.2.0
python-engineio==4.3.3
python-socketio==5.7.1
simple-websocket==0.7.0 
cv2
numpy
base64
gunicorn
